IntelliCVClassifier/modules/intelliCameraCapture/requirements.txt
azure-iot-device~=2.0.0
azure.storage.blob==2.1.0
requests
progress